requirements.txt
alembic==1.11.3
argon2==0.1.10
argon2-cffi==23.1.0
Flask==3.0.3
httpx==0.27.2
PyMySQL==1.0.2
python-dotenv==1.0.1
pytz==2022.1
SQLAlchemy==1.3.0